Get rid of timezone commands
Keep the timezone files because it would be nice
to have the timezone files so that the developers can easily debug, and
also in case anyone wants to see the different timezones in those
file formats.

Also keep the timezone file parsing logic because it defaults to
US/Pacific already, which is really what I want because that makes
things easy to work with. And this is a lot of good adaptible code that
I don't necessarily want to get rid of. It adds a bit of overhead in
terms of script complexity, but there's no performance impact because
the file should only be read once upon startup.